added SwitchDemo and linked the switch functionality
