+++ src/components/ContentComponent.jsx
@@ -13,7 +13,8 @@ import GenerateOTP from "../lib/inputForm/GenerateOTP";
 import InputOTP from "../lib/inputForm/InputOTP";
 import ToDoList from "../lib/inputForm/ToDoList";
 import DragAndDrop from "../lib/drag-and-drop/Area";
-import Switch from "../lib/switch/Switch";
+// import Switch from "../lib/switch/Switch";
+import SwitchDemo from "../lib/switch/SwitchDemo";
 
 export default function ContentComponent(props) {
   //initialize an install step which will dynamically change instructions based on prop comp
@@ -88,7 +89,7 @@ export default function ContentComponent(props) {
             <ToDoList />
           </Show>
           <Show when={currentComp() === "Switch"}>
-            <Switch />
+            <SwitchDemo />
           </Show>
           <Show when={currentComp() === "Drag And Drop"}>
             <DragAndDrop items={2} columns={4} showNewColBtn={false} />
